+++ internal/component/pyroscope/scrape/scrape_test.go
@@ -151,6 +151,16 @@ func TestUnmarshalConfig(t *testing.T) {
 			`,
 			expectedErr: "scrape_interval must be at least 2 seconds when using delta profiling",
 		},
+		"invalid cpu profiling_duration": {
+			in: `
+			targets    = []
+			forward_to = null
+			scrape_timeout = "1s"
+			scrape_interval = "10s"
+			profiling_duration = "1s"
+			`,
+			expectedErr: "profiling_duration must be larger then 1 second when using delta profiling",
+		},
 		"allow short scrape_intervals without delta": {
 			in: `
 			targets    = []