+++ internal/component/pyroscope/scrape/target_test.go
@@ -4,6 +4,7 @@ import (
 	"net/url"
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/discovery/targetgroup"
@@ -230,6 +231,68 @@ func Test_targetsFromGroup(t *testing.T) {
 			}),
 			url.Values{"seconds": []string{"14"}}),
 	}
+	require.NoError(t, err)
+	sort.Sort(Targets(active))
+	sort.Sort(Targets(expected))
+	require.Equal(t, expected, active)
+	require.Empty(t, dropped)
+}
+
+func Test_targetsFromGroup_withSpecifiedProfilingDuration(t *testing.T) {
+	args := NewDefaultArguments()
+	args.ProfilingConfig.Block.Enabled = false
+	args.ProfilingConfig.Goroutine.Enabled = false
+	args.ProfilingConfig.Mutex.Enabled = false
+	args.ProfilingDuration = 20 * time.Second
+
+	active, dropped, err := targetsFromGroup(&targetgroup.Group{
+		Targets: []model.LabelSet{
+			{model.AddressLabel: "localhost:9090"},
+		},
+		Labels: model.LabelSet{
+			"foo": "bar",
+		},
+	}, args, args.ProfilingConfig.AllTargets())
+	expected := []*Target{
+		// unspecified
+		NewTarget(
+			labels.FromMap(map[string]string{
+				model.AddressLabel:    "localhost:9090",
+				serviceNameLabel:      "unspecified",
+				model.MetricNameLabel: pprofMemory,
+				ProfilePath:           "/debug/pprof/allocs",
+				model.SchemeLabel:     "http",
+				"foo":                 "bar",
+				"instance":            "localhost:9090",
+			}),
+			labels.FromMap(map[string]string{
+				model.AddressLabel:    "localhost:9090",
+				model.MetricNameLabel: pprofMemory,
+				ProfilePath:           "/debug/pprof/allocs",
+				model.SchemeLabel:     "http",
+				"foo":                 "bar",
+			}),
+			url.Values{}),
+		NewTarget(
+			labels.FromMap(map[string]string{
+				model.AddressLabel:    "localhost:9090",
+				serviceNameLabel:      "unspecified",
+				model.MetricNameLabel: pprofProcessCPU,
+				ProfilePath:           "/debug/pprof/profile",
+				model.SchemeLabel:     "http",
+				"foo":                 "bar",
+				"instance":            "localhost:9090",
+			}),
+			labels.FromMap(map[string]string{
+				model.AddressLabel:    "localhost:9090",
+				model.MetricNameLabel: pprofProcessCPU,
+				ProfilePath:           "/debug/pprof/profile",
+				model.SchemeLabel:     "http",
+				"foo":                 "bar",
+			}),
+			url.Values{"seconds": []string{"20"}}),
+	}
+
 	require.NoError(t, err)
 	sort.Sort(Targets(active))
 	sort.Sort(Targets(expected))